Implementation of sample profiling
This is a first implementation of sample profiling. For each new case,
with a snv vcf given, a profile for each sample will be made, that is
stored in the case collection, for each individual. Upon uploading new
cases, the profiles of the loaded individuals will be compared against
all profiles in the database. If there already exists similar profiles
the loading will be aborted.